fix:Arreglado double --> int

+++ src/main/java/ejercicios/Ejercicio012.java
@@ -5,8 +5,8 @@ import java.util.Scanner;
 public class Ejercicio012 {
 	
 	//COMPLETAR METODO
-	public double sumarHastaCero() {
-		return 0.0;
+	public int sumarHastaCero() {
+		return 0;
     }
